[xy-chart][tests] fix tests, mirror src dir structure, add tests for eventTrigger, eventTriggerRefs, and new utils
