+++ packages/xy-chart/src/utils/shallowCompareObjectEntries.js
@@ -2,5 +2,5 @@ export default function shallowCompareObjectEntries(a, b) {
   const keysA = Object.keys(a);
   const keysB = Object.keys(b);
   if (keysA.length !== keysB.length) return false;
-  return keysA.some(k => a[k] !== b[k]);
+  return keysA.every(k => a[k] === b[k]);
 }
