+++ packages/xy-chart/test/series/GroupedBarSeries.test.js
@@ -2,7 +2,7 @@ import Bar from '@vx/shape/build/shapes/Bar';
 import BarGroup from '@vx/shape/build/shapes/BarGroup';
 import React from 'react';
 import { shallow, mount } from 'enzyme';
-import { XYChart, GroupedBarSeries } from '../src/';
+import { XYChart, GroupedBarSeries } from '../../src/';
 
 describe('<GroupedBarSeries />', () => {
   const mockProps = {
