+++ packages/xy-chart/test/series/StackedBarSeries.test.js
@@ -2,9 +2,9 @@ import Bar from '@vx/shape/build/shapes/Bar';
 import BarStack from '@vx/shape/build/shapes/BarStack';
 import React from 'react';
 import { shallow, mount } from 'enzyme';
-import { XYChart, StackedBarSeries } from '../src/';
+import { XYChart, StackedBarSeries } from '../../src/';
 
-describe('<GroupedBarSeries />', () => {
+describe('<StackedBarSeries />', () => {
   const mockProps = {
     xScale: { type: 'band' },
     yScale: { type: 'linear', includeZero: false },
